fix: download substrate on app start
