+++ packages/electron-app/src/main/index.ts
@@ -3,105 +3,16 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import electron from 'electron';
-import installExtension, { REACT_DEVELOPER_TOOLS } from 'electron-devtools-installer';
-import Pino from 'pino';
-import path from 'path';
-import url from 'url';
 
-import { CSP, staticPath } from './util';
 import SluiApp from './app';
 
-const { app, BrowserWindow, session } = electron;
-const pino = Pino();
-let mainWindow: Electron.BrowserWindow | undefined;
-
-// https://electronjs.org/docs/tutorial/security#electron-security-warnings
-process.env.ELECTRON_ENABLE_SECURITY_WARNINGS = 'true';
-
-let sluiApp;
+const { app } = electron;
 
 app.once('ready', () => {
-  sluiApp = new SluiApp(app, options);
+  let sluiApp = new SluiApp(app);
 
   return sluiApp;
 });
-
-function createWindow () {
-  mainWindow = new BrowserWindow({
-    height: 1920,
-    resizable: true,
-    width: 1440,
-    webPreferences: {
-      allowRunningInsecureContent: false,
-      contextIsolation: true, // prevent context sharing between renderer<->main processes
-      devTools: true,
-      enableBlinkFeatures: '', // https://electronjs.org/docs/tutorial/security#9-do-not-use-enableblinkfeatures
-      enableRemoteModule: false,
-      experimentalFeatures: false, // Do not set to true
-      navigateOnDragDrop: false,
-      nativeWindowOpen: true,
-      nodeIntegration: false, // https://electronjs.org/docs/tutorial/security#2-disable-nodejs-integration-for-remote-content
-      nodeIntegrationInWorker: false, // https://electronjs.org/docs/tutorial/security#2-disable-nodejs-integration-for-remote-content
-      plugins: false,
-      sandbox: true, // isolate all BrowserWindow instance environments
-      safeDialogs: true,
-      safeDialogsMessage: 'Electron consecutive dialog protection was triggered',
-      webSecurity: true, // https://electronjs.org/docs/tutorial/security#5-do-not-disable-websecurity
-      webviewTag: false // Associated with `will-attach-webview`
-    }
-  });
-
-  mainWindow.loadURL(
-    (process.env.NODE_ENV !== 'production' && process.env.ELECTRON_START_URL) ||
-    url.format({
-      pathname: path.join(staticPath, 'build', 'index.html'),
-      protocol: 'file:',
-      slashes: true
-    })
-  );
-
-  if (process.env.NODE_ENV !== 'production') {
-    installExtension(REACT_DEVELOPER_TOOLS)
-      .then((name: string) => console.log(`Added Extension:  ${name}`))
-      .catch((err: string) => console.log('An error occurred: ', err));
-  }
-
-  // Content Security Policy (CSP)
-  session.defaultSession!.webRequest.onHeadersReceived((details, callback) => {
-    // Note: `onHeadersReceived` will not be called in prod, because we use the
-    // file:// protocol: https://electronjs.org/docs/tutorial/security#csp-meta-tag
-    // Instead, the CSP are the ones in the meta tag inside index.html
-    pino.debug(
-      'Configuring Content-Security-Policy for environment development'
-    );
-
-    callback({
-      responseHeaders: {
-        ...details.responseHeaders,
-        'Content-Security-Policy': [CSP]
-      }
-    });
-  });
-
-  mainWindow.on('closed', () => {
-    mainWindow = undefined;
-  });
-}
-
-app.on('activate', function () {
-  // On OS X it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (mainWindow === undefined) {
-    createWindow();
-  }
-});
-
-// Uncomment this block if ever we decide to use a webview in the app.
-// app.on('web-contents-created', (event, contents) => {
-//   contents.on('will-attach-webview', (event, webPreferences, params) => {
-//     event.preventDefault();
-//   });
-// });
 
 // Quit when all windows are closed.
 app.on('window-all-closed', function () {
@@ -111,3 +22,10 @@ app.on('window-all-closed', function () {
     app.quit();
   }
 });
+
+// Uncomment this block if ever we decide to use a webview in the app.
+// app.on('web-contents-created', (event, contents) => {
+//   contents.on('will-attach-webview', (event, webPreferences, params) => {
+//     event.preventDefault();
+//   });
+// });
